Move handle_request to ChainWorkerState.
