+++ linera-core/src/chain_worker/state/mod.rs
@@ -29,6 +29,7 @@ use linera_execution::{ExecutionStateView, Query, QueryOutcome, ServiceRuntimeEn
 use linera_storage::{Clock as _, ResultReadCertificates, Storage};
 use linera_views::views::ClonableView;
 use tokio::sync::{oneshot, OwnedRwLockReadGuard, RwLock};
+use tracing::{instrument, warn};
 
 #[cfg(test)]
 pub(crate) use self::attempted_changes::CrossChainUpdateHelper;
@@ -36,7 +37,7 @@ use self::{
     attempted_changes::ChainWorkerStateWithAttemptedChanges,
     temporary_changes::ChainWorkerStateWithTemporaryChanges,
 };
-use super::{ChainWorkerConfig, DeliveryNotifier};
+use super::{ChainWorkerConfig, ChainWorkerRequest, DeliveryNotifier};
 use crate::{
     data_types::{ChainInfoQuery, ChainInfoResponse, CrossChainRequest},
     value_cache::ValueCache,
@@ -97,6 +98,118 @@ where
     /// Returns the [`ChainId`] of the chain handled by this worker.
     pub fn chain_id(&self) -> ChainId {
         self.chain.chain_id()
+    }
+
+    /// Runs the worker until there are no more incoming requests.
+    #[instrument(skip(self, request))]
+    pub async fn handle_request(&mut self, request: ChainWorkerRequest<StorageClient::Context>) {
+        // TODO(#2237): Spawn concurrent tasks for read-only operations
+        let responded = match request {
+            #[cfg(with_testing)]
+            ChainWorkerRequest::ReadCertificate { height, callback } => {
+                callback.send(self.read_certificate(height).await).is_ok()
+            }
+            #[cfg(with_testing)]
+            ChainWorkerRequest::FindBundleInInbox {
+                inbox_id,
+                certificate_hash,
+                height,
+                index,
+                callback,
+            } => callback
+                .send(
+                    self.find_bundle_in_inbox(inbox_id, certificate_hash, height, index)
+                        .await,
+                )
+                .is_ok(),
+            ChainWorkerRequest::GetChainStateView { callback } => {
+                callback.send(self.chain_state_view().await).is_ok()
+            }
+            ChainWorkerRequest::QueryApplication { query, callback } => {
+                callback.send(self.query_application(query).await).is_ok()
+            }
+            ChainWorkerRequest::DescribeApplication {
+                application_id,
+                callback,
+            } => callback
+                .send(self.describe_application(application_id).await)
+                .is_ok(),
+            ChainWorkerRequest::StageBlockExecution {
+                block,
+                round,
+                published_blobs,
+                callback,
+            } => callback
+                .send(
+                    self.stage_block_execution(block, round, &published_blobs)
+                        .await,
+                )
+                .is_ok(),
+            ChainWorkerRequest::ProcessTimeout {
+                certificate,
+                callback,
+            } => callback
+                .send(self.process_timeout(certificate).await)
+                .is_ok(),
+            ChainWorkerRequest::HandleBlockProposal { proposal, callback } => callback
+                .send(self.handle_block_proposal(proposal).await)
+                .is_ok(),
+            ChainWorkerRequest::ProcessValidatedBlock {
+                certificate,
+                callback,
+            } => callback
+                .send(self.process_validated_block(certificate).await)
+                .is_ok(),
+            ChainWorkerRequest::ProcessConfirmedBlock {
+                certificate,
+                notify_when_messages_are_delivered,
+                callback,
+            } => callback
+                .send(
+                    self.process_confirmed_block(certificate, notify_when_messages_are_delivered)
+                        .await,
+                )
+                .is_ok(),
+            ChainWorkerRequest::ProcessCrossChainUpdate {
+                origin,
+                bundles,
+                callback,
+            } => callback
+                .send(self.process_cross_chain_update(origin, bundles).await)
+                .is_ok(),
+            ChainWorkerRequest::ConfirmUpdatedRecipient {
+                recipient,
+                latest_height,
+                callback,
+            } => callback
+                .send(
+                    self.confirm_updated_recipient(recipient, latest_height)
+                        .await,
+                )
+                .is_ok(),
+            ChainWorkerRequest::HandleChainInfoQuery { query, callback } => callback
+                .send(self.handle_chain_info_query(query).await)
+                .is_ok(),
+            ChainWorkerRequest::DownloadPendingBlob { blob_id, callback } => callback
+                .send(self.download_pending_blob(blob_id).await)
+                .is_ok(),
+            ChainWorkerRequest::HandlePendingBlob { blob, callback } => {
+                callback.send(self.handle_pending_blob(blob).await).is_ok()
+            }
+            ChainWorkerRequest::UpdateReceivedCertificateTrackers {
+                new_trackers,
+                callback,
+            } => callback
+                .send(
+                    self.update_received_certificate_trackers(new_trackers)
+                        .await,
+                )
+                .is_ok(),
+        };
+
+        if !responded {
+            warn!("Callback for `ChainWorkerActor` was dropped before a response was sent");
+        }
     }
 
     /// Returns a read-only view of the [`ChainStateView`].
